+++ src/Router.tsx
@@ -1,7 +1,7 @@
 import { createBrowserRouter, Navigate, RouterProvider } from 'react-router-dom';
 import { lazy } from 'react';
 import { Home } from './pages/Home';
-import { AppLayout, ContentLayout } from './features/layout/Layout';
+import { AppLayout, ContentLayout } from './components/layout';
 const Detail = lazy(() => import('./pages/Detail'));
 
 export const Router = () => {
