+++ src/components/layout/index.ts
@@ -1,0 +1,2 @@
+export * from './AppLayout';
+export * from './ContentLayout';
